% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy-season.R
\name{dummySeason}
\alias{dummySeason}
\title{A function to generate dummy inputs}
\usage{
dummySeason(values)
}
\arguments{
\item{values}{the vector of values}
}
\value{
the matrix of the inputs
}
\description{
This function build a matrix of n inputs from a given array of values
}
\examples{
   levels <- sapply(1:10, function(x) sprintf("\%02d", x))

   sample <- sample(levels, 100, replace=TRUE)
   dummySample <- dummySeason(sample);
   head(dummySample)
}
